Updated project structure to be more scalable Cleaned up code

+++ modules/collage.py
@@ -1,8 +1,7 @@
 from PIL import Image
 import os
 
-IMAGE_LIST = os.listdir("images")
-print(sorted(IMAGE_LIST))
+IMAGE_LIST = os.listdir("collage/images")
 IMAGE_WIDTH = 1000
 IMAGE_HEIGHT = 600
 PATH = 'collage/'
@@ -24,22 +23,23 @@ def create_collage(vertical, horizontal):
     for index, i in enumerate(IMAGE_LIST):
         # CREATE NEW CANVAS IF CANVAS REACHES MAX CAPACITY
         if index != 0 and index % total == 0:
-            collage.save(PATH + '{}.png'.format(count))
+            collage.save(PATH + '{}-collage.png'.format(count))
             collage = Image.new("RGBA", size=(width, height))
             count += 1
             x_range = 0
             y_range = 0
 
-        collage.paste(Image.open("images/{}".format(i)), (x_range, y_range))
+        collage.paste(Image.open("collage/images/{}".format(i)), (x_range, y_range))
 
         # MOVE PASTE COORDINATES BY THE SIZE OF ONE IMAGE
         x_range += IMAGE_WIDTH
         if x_range == width:
             y_range += IMAGE_HEIGHT
             x_range = 0
-    collage.save(PATH + '/{}.png'.format(count))
+    collage.save(PATH + '/{}-collage.png'.format(count))
 
 
 def empty_folder():
     for f in os.listdir(PATH):
-        os.remove(os.path.join(PATH, f))
+        if f.endswith('.png'):
+            os.remove(os.path.join(PATH, f))
